# Api/src/docs/openapi-autogenerated.yaml
openapi: 3.0.0
info:
  title: St Michel Website API
  version: 0.1.0
  description: St Michel Website Backend
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /language/{id}:
    delete:
      tags:
        - Language
      security:
        - Authorization: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Language ID
      summary: delete a language
      responses:
        '200':
          description: Language deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Language deleted successfully
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid credentials
        '500':
          description: Internal Server error
    put:
      tags:
        - Language
      security:
        - Authorization: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Language ID
      summary: Update a Mass language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  example: St Michael Parish
                  description: language of the Mass
                isActive:
                  type: boolean
                  example: 'false'
                  description: Activation state of the language
      responses:
        '200':
          description: Language saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Language updated successfully.
        '401':
          description: Invalid credentials!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid credentials!
        '404':
          description: Language not found!
        '422':
          description: Validation error!
        '500':
          description: Internal Server error!
...
  /mass-times/{id}:
    put:
      tags:
        - Mass Times
      security:
        - Authorization: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Mass time ID
      summary: Update a Mass Time
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  type: number
                  example: 1
                  description: location of the Mass
                language:
                  type: number
                  example: 1
                  description: language of the Mass
                day_en:
                  type: string
                  example: Sunday 07 July 2024
                  description: day of the Mass
                day_rw:
                  type: string
                  example: Sunday 07 July 2024
                  description: day of the Mass
                day_fr:
                  type: string
                  example: Sunday 07 July 2024
                  description: day of the Mass
                time:
                  type: string
                  example: 9 AM
                  description: time of the Mass
                isActive:
                  type: boolean
                  example: 'false'
                  description: Activation state of the location
      responses:
        '200':
          description: Mass Time Update
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Mass Time updated successfully.
        '401':
          description: Invalid credentials!
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid credentials!
        '404':
          description: not found!
        '422':
          description: Validation error!
        '500':
          description: Internal Server error!
...